Fix 컬럼수정 point nullable -> Long or Integer
변경

+++ src/main/java/com/onz/modules/account/web/dto/response/AccountResponse.java
@@ -19,7 +19,7 @@ public class AccountResponse {
     private String userId;
     private Role role;
     private AuthProvider snsType;
-    private long point;
+    private Integer point;
     private Gubn gubn;
     private Myinfo myinfo;
     private YN isDelete;
